Building out the basic struture of the app
The app has 2 main web components -- the app and an individual list item.

The app does not currently work. When you try to add a new todo item you'll get a "Cannot read property 'ownerDocument' of null" error because `this._doc.currentScript` is only valid at initialization. This property needs to be cached on the prototype for future access.